moved dtos and interfaces to orchestrator contract
