+++ Orchestrator/Prognetics.Saga.Orchestrator.Contract/DTO/EngineInput.cs
@@ -1,6 +1,4 @@
-using Prognetics.Saga.Orchestrator.Contract.DTO;
-
-namespace Prognetics.Saga.Orchestrator;
+namespace Prognetics.Saga.Orchestrator.Contract.DTO;
 
 public readonly record struct EngineInput(
     string EventName,
