+++ Orchestrator/Prognetics.Saga.Orchestrator.Contract/ISagaEngine.cs
@@ -1,6 +1,6 @@
 using Prognetics.Saga.Orchestrator.Contract.DTO;
 
-namespace Prognetics.Saga.Orchestrator;
+namespace Prognetics.Saga.Orchestrator.Contract;
 public interface ISagaEngine
 {
     Task<EngineOutput?> Process(EngineInput input);
