--- Orchestrator/Prognetics.Saga.Orchestrator/ICompensationStore.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace Prognetics.Saga.Orchestrator;
-
-public interface ICompensationStore{
-    Task<IEnumerable<Compensation>> GetCompensations(string transactionId);
-    Task SaveCompensation(Compensation compensation);
-}
